MWPW-172007: fix test
