--- eds/blocks/knowledge-base-overview/KnowledgeBaseOverview.js
@@ -1,3 +1,0 @@
-import PartnerCardsWithDateFilter from '../../components/PartnerCardsWithDateFilter.js';
-
-export default class KnowledgeBaseOverview extends PartnerCardsWithDateFilter {}
