--- eds/blocks/partner-news/PartnerNews.js
@@ -1,3 +1,0 @@
-import PartnerCardsWithDateFilter from '../../components/PartnerCardsWithDateFilter.js';
-
-export default class PartnerNews extends PartnerCardsWithDateFilter {}
