+++ test/blocks/dx-card-collection/mocks/body.html
@@ -1,4 +1,4 @@
-<div class="knowledge-base-overview">
+<div class="dx-card-collection">
   <div>
     <div>Title</div>
     <div>Sample Title for Card Collection</div>
