+++ test/scripts/mocks/dx-card-collection.html
@@ -1,4 +1,4 @@
-<div class="partner-news">
+<div class="dx-card-collection">
   <div>
     <div>Title</div>
     <div>Sample Title for Card Collection</div>
@@ -10,6 +10,16 @@
     <div></div>
   </div>
   <div>
+    <div data-valign="middle">Collection Tags</div>
+    <div data-valign="middle">
+      <ul>
+        <li>caas:adobe-partners/qa-content</li>
+      </ul>
+    </div>
+    <div data-valign="middle"></div>
+  </div>
+  <div>
+
     <div>Filter</div>
     <div>
       <ul>
