--- test/scripts/mocks/partner-news.html
@@ -1,120 +1,0 @@
-<div class="partner-news">
-  <div>
-    <div>Title</div>
-    <div>Sample Title for Card Collection</div>
-    <div></div>
-  </div>
-  <div>
-    <div>Cards per page</div>
-    <div>6</div>
-    <div></div>
-  </div>
-  <div>
-    <div data-valign="middle">Collection Tags</div>
-    <div data-valign="middle">
-      <ul>
-        <li>caas:adobe-partners/qa-content</li>
-      </ul>
-    </div>
-    <div data-valign="middle"></div>
-  </div>
-  <div>
-
-    <div>Filter</div>
-    <div>
-      <ul>
-        <li>Applications</li>
-      </ul>
-    </div>
-    <div>
-      <ul>
-        <li>sign</li>
-        <li>advertising-cloud</li>
-        <li>analytics</li>
-        <li>audience-manager</li>
-        <li>campaign</li>
-        <li>commerce</li>
-        <li>customer-journey-analytics</li>
-        <li>document-cloud</li>
-        <li>document-services</li>
-        <li>experience-cloud</li>
-        <li>experience-manager</li>
-        <li>experience-platform</li>
-        <li>journey-optimizer</li>
-        <li>marketo-engage</li>
-        <li>microsoft</li>
-        <li>real-time-cdp</li>
-        <li>target</li>
-        <li>workfront</li>
-      </ul>
-    </div>
-  </div>
-  <div>
-    <div>Filter</div>
-    <div>
-      <ul>
-        <li>audience</li>
-      </ul>
-    </div>
-    <div>
-      <ul>
-        <li>marketing</li>
-        <li>sales</li>
-        <li>practice-lead</li>
-        <li>technical</li>
-      </ul>
-    </div>
-  </div>
-  <div>
-    <div>Filter</div>
-    <div>
-      <ul>
-        <li>region</li>
-      </ul>
-    </div>
-    <div>
-      <ul>
-        <li>americas</li>
-        <li>asia-pacific</li>
-        <li>emea</li>
-        <li>japan</li>
-      </ul>
-    </div>
-  </div>
-  <div>
-    <div>Filter</div>
-    <div>
-      <ul>
-        <li>topic</li>
-      </ul>
-    </div>
-    <div>
-      <ul>
-        <li>analyst-reports</li>
-        <li>amplify</li>
-        <li>certification-credentials</li>
-        <li>enablement-training</li>
-        <li>events</li>
-        <li>executive-insights</li>
-        <li>industries</li>
-        <li>marketing</li>
-        <li>program-benefits</li>
-        <li>programs-contests</li>
-        <li>product-releases</li>
-        <li>selling</li>
-        <li>solutions</li>
-        <li>specialization</li>
-      </ul>
-    </div>
-  </div>
-  <div>
-    <div>Sort</div>
-    <div>
-      <ul>
-        <li>newest</li>
-        <li>oldest_default</li>
-      </ul>
-    </div>
-    <div></div>
-  </div>
-</div>
